app: move Wake* to app.internal.wake
We need to change the cleanup logic a bit, as we cannot access the
wake_session variable from within the app.internal.wake module.
